+++ changelog.md
@@ -1,3 +1,10 @@
+0.2.2
+=====
+
+Now partially using DOM methods to create a DOM document instead of using string
+serialization for the `xpath()` method.
+
+
 0.2.1
 =====